refactor(WalletForm): extract topup card
